+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added the ability to color histograms
 - added a property for rounding values on the graph
 - added a tooltip with values on columns
+- click event when clicking on the item
 
 ### Fixed
 - fixed errors in the console when leaving the dashboard